net-test: packetdrill: Add EDO Extension option.
Once EDO is negotiated successfully with the EDO Supported option
in SYN and SYN+ACK, we can include the EDO Extension option in the
subsequent segments to use more than 40 bytes for TCP options.

The EDO Extension option has two variants.  Both variants have TCP
header length that overwrites doff, and the longer variant has the
segment length of the TCP pseudo-header.

We can express these options as follows:

  * <edo> : Longer variant with auto-completion
  * <edo INTEGER> : Simple variant
  * <edo INTEGER INTEGER> : Longer variant

  Auto completion is enabled for inbound packets If INTEGER is -1.

Also, the options are included in the following format:

  Simple variant:

  +--------+--------+--------+--------+
  |  Kind  | Length |  ExID (0x0ED0)  |   Kind   = 254 (Experimental)
  +--------+--------+--------+--------+   Length =   6
  |  Header_Length  |
  +--------+--------+

  Longer variant:

  +--------+--------+--------+--------+
  |  Kind  | Length |  ExID (0x0ED0)  |   Kind   = 254 (Experimental)
  +--------+--------+--------+--------+   Length =   8
  |  Header_Length  |  Segment_Length |
  +--------+--------+--------+--------+

Note that this commit does not extend the TCP options area, and we
will do it in the following commit.

Signed-off-by: Kuniyuki Iwashima <kuniyu@amazon.com>