net-test: packetdrill: Move offset_ptr() to types.h.
We will support the EDO Extension option and auto-completion for the
two length fields.

If auto-completion is used, we save the position of the EDO Extension
option in struct tcp_options and fill the header and segment length
just before sending the packet.

When we re-allocate memory for TCP options, we need to map the position
in the old buffer to the new one with offset_ptr().

So, let's export offset_ptr() to a more general place.

Signed-off-by: Kuniyuki Iwashima <kuniyu@amazon.com>